refactor: replace GST Sales Register with GST Sales Register Beta
