+++ india_compliance/gst_india/report/gst_sales_register/gst_sales_register.json
@@ -1,15 +1,15 @@
 {
  "add_total_row": 1,
  "columns": [],
- "creation": "2017-06-21 16:44:41.621260",
- "disable_prepared_report": 0,
+ "creation": "2024-03-11 16:02:41.778873",
  "disabled": 0,
  "docstatus": 0,
  "doctype": "Report",
  "filters": [],
  "idx": 0,
  "is_standard": "Yes",
- "modified": "2022-03-01 15:45:28.168184",
+ "letterhead": null,
+ "modified": "2024-03-11 16:03:18.740603",
  "modified_by": "Administrator",
  "module": "GST India",
  "name": "GST Sales Register",
