--- india_compliance/gst_india/report/gst_sales_register_beta/gst_sales_register_beta.json
@@ -1,29 +1,0 @@
-{
- "add_total_row": 1,
- "columns": [],
- "creation": "2024-03-11 16:02:41.778873",
- "disabled": 0,
- "docstatus": 0,
- "doctype": "Report",
- "filters": [],
- "idx": 0,
- "is_standard": "Yes",
- "letterhead": null,
- "modified": "2024-03-11 16:03:18.740603",
- "modified_by": "Administrator",
- "module": "GST India",
- "name": "GST Sales Register Beta",
- "owner": "Administrator",
- "prepared_report": 0,
- "ref_doctype": "Sales Invoice",
- "report_name": "GST Sales Register Beta",
- "report_type": "Script Report",
- "roles": [
-  {
-   "role": "Accounts Manager"
-  },
-  {
-   "role": "Accounts User"
-  }
- ]
-}
